docker/srcs/django/requirements.txt
#To install specific libs
django
psycopg2
django-allauth
requests